CMake: Handle multiple libraries being returned for Brotli
Brotli is built as three libraries: libbrotlienc, libbrotlidec, and
libbrotlicommon. When requesting the linker flags using pkg-config for
e.g. libbrotlidec, it will return both libbrotlidec and libbrotlicommon.
The FindBrotli*.cmake files ignore the names of the libraries returned
by pkg-config, and hardcode only the libbrotli{enc,dec} names. While
this is fine when using shared libraries (they contain an entry for
libbrotlicommon as required library in their headers), it will cause
linker errors when Brotli has been built as static libraries, due to
the missing symbols from libbrotlicommon being unknown to the linker.

This fixes FindBrotli*.cmake files by using the library names reported
by pkg-config, and applying find_library() to each of the libraries
to find their absolute paths. The well-known names of the libraries
are still used as fallback in case pkg-config is not available or in
case it does not provide the names. If any of the libraries is
missing, the corresponding BROTLI{ENC,DEC}_LIBRARIES is unset to
let find_package_handle_standard_args() report an error.